feat: Add test data generation script for performance metrics
- Created generate_test_performance_data.py script
- Generate realistic performance data without API costs
- 5 scenarios: realistic, varied, competitive, bob-dominates, trending
- Generates scores, categories, ratings, costs, timestamps
- Reset function to clear all data
- Documented in scripts/README.md

Benefits:
- Test leaderboards with 800+ tasks in ~2 seconds
- Save -5 per test run (no LLM calls)
- Different scenarios for edge case testing
- Easy reset and regenerate

Bug Fixes:
- Fixed Decimal JSON serialization in stats_aggregation.py
- Convert category_performance and model_performance avg_scores to float
- Prevents 'Decimal is not JSON serializable' error

Usage:
  python scripts/generate_test_performance_data.py --tasks 50 --reset
  python scripts/generate_test_performance_data.py --scenario varied --tasks 100

+++ backend/jobs/stats_aggregation.py
@@ -209,7 +209,7 @@ class StatsAggregationJob:
         for category, score_list in categories.items():
             category_stats[category] = {
                 "count": len(score_list),
-                "avg_score": round(sum(score_list) / len(score_list), 2)
+                "avg_score": float(round(sum(score_list) / len(score_list), 2))
             }
 
         return category_stats
@@ -250,7 +250,7 @@ class StatsAggregationJob:
         for model_name, data in models.items():
             model_stats[model_name] = {
                 "count": len(data["scores"]),
-                "avg_score": round(sum(data["scores"]) / len(data["scores"]), 2),
+                "avg_score": float(round(sum(data["scores"]) / len(data["scores"]), 2)),
                 "since": data["first_used"].isoformat() if data["first_used"] else None
             }
